Initial Implementation of the QR Login Command
Note: This will replace the login command once development is complete.

The command takes a device codeURL, tokenURL and clientID of the authentication
service. The command displays a url and qr prompt for the user to login. Once
this login is performed, an id token, access token and refresh token are
downloaded from the authentication service.

ID Token - Provides information about the logged in user
Access Token - Is subsequently passed to the api service to allow access
Refresh Token - Used to refresh the access token once it expires

In the future the IVCAP service will return the code, token and client id
directly rather than the user providing them.