Query the IVCAP service for login URLs
This updates the qrlogin command to first query the IVCAP "/logininfo"
endpoint for urls it should use for login. These are returned as a JSON object
and stored in the AuthInfo struct. The user now only has to provide the address
of the IVCAP service to the command line tool rather than three seperate
login endpoints.

Note: Requires a new deployment of the IVCAP service which has the endpoint