Update login to use updated YAML authinfo configuration file
This updates the login command to use the updated format of the authinfo yaml
configuration file. This file now can return multiple identity providers,
however at the moment will only use the default provider. It also now
provides a version number, so that configuration structure updates can be
handled cleanly in the future.