cmd/login: Replace direct http calls with adapter API and fixed some clock drift issues
To reduce code we need to maintain, all direct http interactions are now going through the `adapter` API.
Also moved duplicated code into their own functions.
In addition, there was a subtle timing bug when login in and immediately checking the token. In case of
an "ErrTokenUsedBeforeIssued' error we now simply wait a bit and try again. We aren't checking on
how often we do that, though, assuming this is only temporary.
There was also an issue with parsing the CustomIdClaims. The account ID is actually returned as UUID
in 'acc'.