Improve handling of access-token
Moved most o the logic in obtaining the access token into login.go/getAccessToken. This function now first checks the respective
env and flag variables and immediately returns those. Only then will it check the active context. An additional 'refreshExpired' parameter
allows us to query the current status without unnecessary refresh procedure (used in IsAuthorised).

We now also clarify the Authorisation status in `context get`. If the token is provided we now state that we have a token, but we don't
know anything about it.

Finally, the default for `context get` is `all` and with it the table.