Fix the command name, and version metadata, of the command installed by 'go install'
The first issue was that the command installed by `make install` was `ivcap-cli`, but it should be `ivcap`.
This is because `go install` doesn't allow you to rename the binary -- it will always be the name of the
directory containing the main module of the command being installed, in this case `ivcap-cli`.
The fix to this issue was to make an `ivcap` subdirectory and put the main package in there.

The next issue was that the version information inserted using the `-ldflags ...` command to
 `go build` was not making it into the installed binary. It turns out that `go install` doesn't
use the binary created by `go build`, it compiles it itself! The fix in this case was to pass
the same `-ldflags ...` command to `go install`.

Note that the `install` make target now no longer depends on the `build` target, because it in
fact does not use the binary that the `build` target creates.