feat(adapter): refactor URL parsing
This commit refactors the URL parsing and error handling in the
Connect function of the adapter package. This refacoring was
undertaken to correct an issue (#38) where creating a context
with a URL containing a trailing backslash results in malformed
requests.

The 'path' parameter is renamed to 'endpoint' to better reflect
its purpose. A new function, parseURL, is introduced to handle
the parsing of the endpoint and the resolution against the base
URL. This function also checks if the scheme of the URL is 'http'.
The error handling is improved by returning more specific errors.
Unit tests are added for the new parseURL function.