refactor(tests): separate setup from service tests
This commit separates the setup code from the service tests into a new
file, setup_test.go. This makes the tests more modular and easier to
maintain. The setup code includes the creation of the adapter, logger,
and test token. The service tests now only contain the test functions
and related variables.